Added two new scripts to better reproduce figure 3C in Vervaeke2012.
Koen clarified via email that the cell being tested was 210710_C1,
that the attenuation was measured over an apical dendrite (one that he
could have pathced during the experiments), and that it wasn't
averaged over different dendrites because the experimental data is
single-trial as well. Also, he specified that the experimental
protocol consisted in stimulating at a distance from the soma and then
measuring the attenuation of the voltage response going down along the
dendrite towards the soma.