Making larger changes in morphology beyond the apical dendrites.
I managed to get the dendritic attenuation length to match between the
models, but the absolute response properties of the cell to synaptic
input are still quite different (Koen's model is more reactive to
dendritic stimulation). This is probably due to a smaller soma,
thinner dendrites and the absence of an axon. At the moment, I'm still
experimenting.