Fixed 'double spike' bug caused by on old KBK channel implementation.
This is a well-known problem with an old version of the mod file for
the KBK channel, and had been squashed somewhere else (at least in the
original model by Solinas - see relevant files on ModelDB and Open
Source Brain - and some of the cells in the original Vervaeke2010
project) in the past, but somehow ended up hiding in here. Probably,
Koen imported the wrong version of the mechanism to be used for
210710_C1 in his 2012 project.

Basically, the TABLE statement needs to be taken out, and the maximum
conductance for the channel needs to be adjusted. This bug had a major
effect on the spiking pattern of the cell, and now I have to
re-calibrate pretty much everythin in my new reduced morphology
model (even though my model, being in pure NeuroML, is not directly
affected).