Fixed missing GJ netconn and generally improved GJs for reduced model.
Now the first segment of the apical dendrites in the reduced model is
considered as being still in the molecular layer. This is to better
reproduce the choices made for the detailed model, and it plays an
important part in gap junction connectivity. As a consequence, the
four GJ connections are organised in the following way:

'GCL' GJs -> basolateral dends and first compartment of apical dends
'MF1' GJs -> second compartment of apical dends
'MF2' and 'MF3' GJs -> third compartment of apical dends

Note how this changes the network I/O relationships.