Added network (de)synchronisation analysis as in Vervaeke2010.
Two new simulation configurations, network_desynchronisation and
network_desynchronisation_2010gap, try to reproduce the
desynchronisation effect reported in Vervaeke2010. The expected
behaviour seems to emerge only if the gap junction connectivity is
implemented as is Vervaeke2010, but not if it is implemented as in
Vervaeke2012: in this case, the network just stays completely
synchronised if one excites 10 cells out of 45. It might desynchronise
if more are excited, but I sill need to check that.

Note that at this stage there is no heterogeneity in the cells' leak
conductance across the network, but there is stochasticity in the
background noise that all cells receive. Thus, these simulations are
kind of halfway between figure 8A (heterogeneous network,
heterogeneous cells, stochastic noise) and figure 9B (heterogeneous
network, homogeneous cells, no noise) in Vervaeke2010.